fix : modified create and getAll services / feat : add update service

+++ logisticsCompany/src/main/java/com/logisticsCompany/repository/PackageEntityRepository.java
@@ -1,12 +1,10 @@
 package com.logisticsCompany.repository;
 
-import com.logisticsCompany.dto.PackageReponseDto;
+
 import com.logisticsCompany.entities.PackageEntity;
-import io.micrometer.observation.ObservationFilter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
 
-public interface PackageEntityRepository extends JpaRepository<PackageEntity, Integer> {
+public interface PackageEntityRepository extends JpaRepository<PackageEntity, Long> {
 
 }
